requirements.txt
streamlit
langchain
openai
python-dotenv
tiktoken
faiss-cpu
PyPDF2
streamlit-authenticator
pymongo
certifi
langchain-community
langchain-openai
toml
